4일차/pyproject.toml
[tool.poetry]
name = "3일차"
version = "0.1.0"
description = ""
authors = ["romiii35 <romiii9735@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.11.7"
orjson = "3.10.12"
sqids = "0.5.2"
tortoise-orm = {version = "0.23.0", extras = ["asyncmy"]}
cryptography = "44.0.0"
aerich = "0.8.1"
tomlkit = "0.13.2"
pydantic-settings = "2.7.1"
httpx = "0.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
black = "24.10.0"
ruff = "0.8.2"
mypy = "1.13.0"
coverage = "7.6.9"
pytest-asyncio = "0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"



[tool.aerich]

tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"

location = "./migrations"

src_folder = "./."

